Add JS++ and fix typo on TypeScript
JS++ is a seemingly inspired by C# language which is unfortunately rather unknown, but definitively an alternative to Flow. However, it is different from TypeScript and Flow as it does not support incremental compilation (yet?)

Also, "Typescript" might be right, but per unwritten convention, a scripting language's name including the word "script" has that word capitalized - or rather, camelcased. Hence: Type**S**cript.